+++ tests/test_structure_counter.py
@@ -22,59 +22,76 @@ class TestStructure(BaseTmpl):
 
 class TestStructureCounter(BaseTmpl):
     def setUp(self):
-        self.selected_measures = {"VP", "T", "DC_C"}
-        self.counter = StructureCounter(selected_measures=self.selected_measures)
+        self.selected_measures1 = {"VP", "T", "DC_C"}
+        self.selected_measures2 = {"CP", "VP_T", "CN_C"}
+        self.counter1 = StructureCounter(selected_measures=self.selected_measures1)
+        self.counter2 = StructureCounter(selected_measures=self.selected_measures2)
         return super().setUp()
 
+    def test_parse_selected_measures(self):
+        self.assertListEqual(
+            ["VP", "T", "DC_C"], [s.name for s in self.counter1.structures_to_report]
+        )
+        self.assertListEqual(
+            ["VP1", "VP2", "C1", "T1", "T2", "DC"],
+            [s.name for s in self.counter1.structures_to_query],
+        )
+        self.assertListEqual(
+            ["CP", "VP_T", "CN_C"], [s.name for s in self.counter2.structures_to_report]
+        )
+        self.assertListEqual(
+            ["VP1", "C1", "T1", "T2", "CN1", "CN2", "CN3", "CP"],
+            [s.name for s in self.counter2.structures_to_query],
+        )
     def test_update_freqs(self):
-        self.counter.VP1.freq, self.counter.VP2.freq = 7, 1
-        self.counter.C1.freq, self.counter.C2.freq = 9, 1
-        self.counter.T1.freq, self.counter.T2.freq = 11, 1
-        self.counter.CN1.freq, self.counter.CN2.freq, self.counter.CN3.freq = 13, 1, 2
+        self.counter1.VP1.freq, self.counter1.VP2.freq = 7, 1
+        self.counter1.C1.freq, self.counter1.C2.freq = 9, 1
+        self.counter1.T1.freq, self.counter1.T2.freq = 11, 1
+        self.counter1.CN1.freq, self.counter1.CN2.freq, self.counter1.CN3.freq = 13, 1, 2
 
-        self.counter.update_freqs()
-        self.assertEqual(8, self.counter.VP.freq)
-        self.assertEqual(10, self.counter.C.freq)
-        self.assertEqual(12, self.counter.T.freq)
-        self.assertEqual(16, self.counter.CN.freq)
+        self.counter1.update_freqs()
+        self.assertEqual(8, self.counter1.VP.freq)
+        self.assertEqual(10, self.counter1.C.freq)
+        self.assertEqual(12, self.counter1.T.freq)
+        self.assertEqual(16, self.counter1.CN.freq)
 
     def test_compute_14_indices(self):
         (
-            self.counter.W.freq,
-            self.counter.S.freq,
-            self.counter.VP1.freq,
-            self.counter.C1.freq,
-            self.counter.T.freq,
-            self.counter.DC.freq,
-            self.counter.CT.freq,
-            self.counter.CP.freq,
-            self.counter.CN.freq,
+            self.counter1.W.freq,
+            self.counter1.S.freq,
+            self.counter1.VP1.freq,
+            self.counter1.C1.freq,
+            self.counter1.T.freq,
+            self.counter1.DC.freq,
+            self.counter1.CT.freq,
+            self.counter1.CP.freq,
+            self.counter1.CN.freq,
         ) = (1530, 54, 174, 149, 83, 61, 52, 26, 181)
-        self.counter.compute_14_indicies()
-        self.assertEqual(1530, self.counter.W.freq)
-        self.assertEqual(54, self.counter.S.freq)
-        self.assertEqual(174, self.counter.VP1.freq)
-        self.assertEqual(149, self.counter.C1.freq)
-        self.assertEqual(83, self.counter.T.freq)
-        self.assertEqual(61, self.counter.DC.freq)
-        self.assertEqual(52, self.counter.CT.freq)
-        self.assertEqual(26, self.counter.CP.freq)
-        self.assertEqual(181, self.counter.CN.freq)
-        self.assertEqual(28.3333, self.counter.MLS.freq)
-        self.assertEqual(18.4337, self.counter.MLT.freq)
-        self.assertEqual(10.2685, self.counter.MLC.freq)
-        self.assertEqual(2.7593, self.counter.C_S.freq)
-        self.assertEqual(2.0964, self.counter.VP_T.freq)
-        self.assertEqual(1.7952, self.counter.C_T.freq)
-        self.assertEqual(0.4094, self.counter.DC_C.freq)
-        self.assertEqual(0.7349, self.counter.DC_T.freq)
-        self.assertEqual(1.537, self.counter.T_S.freq)
-        self.assertEqual(0.6265, self.counter.CT_T.freq)
-        self.assertEqual(0.3133, self.counter.CP_T.freq)
-        self.assertEqual(0.1745, self.counter.CP_C.freq)
-        self.assertEqual(2.1807, self.counter.CN_T.freq)
-        self.assertEqual(1.2148, self.counter.CN_C.freq)
+        self.counter1.compute_14_indicies()
+        self.assertEqual(1530, self.counter1.W.freq)
+        self.assertEqual(54, self.counter1.S.freq)
+        self.assertEqual(174, self.counter1.VP1.freq)
+        self.assertEqual(149, self.counter1.C1.freq)
+        self.assertEqual(83, self.counter1.T.freq)
+        self.assertEqual(61, self.counter1.DC.freq)
+        self.assertEqual(52, self.counter1.CT.freq)
+        self.assertEqual(26, self.counter1.CP.freq)
+        self.assertEqual(181, self.counter1.CN.freq)
+        self.assertEqual(28.3333, self.counter1.MLS.freq)
+        self.assertEqual(18.4337, self.counter1.MLT.freq)
+        self.assertEqual(10.2685, self.counter1.MLC.freq)
+        self.assertEqual(2.7593, self.counter1.C_S.freq)
+        self.assertEqual(2.0964, self.counter1.VP_T.freq)
+        self.assertEqual(1.7952, self.counter1.C_T.freq)
+        self.assertEqual(0.4094, self.counter1.DC_C.freq)
+        self.assertEqual(0.7349, self.counter1.DC_T.freq)
+        self.assertEqual(1.537, self.counter1.T_S.freq)
+        self.assertEqual(0.6265, self.counter1.CT_T.freq)
+        self.assertEqual(0.3133, self.counter1.CP_T.freq)
+        self.assertEqual(0.1745, self.counter1.CP_C.freq)
+        self.assertEqual(2.1807, self.counter1.CN_T.freq)
+        self.assertEqual(1.2148, self.counter1.CN_C.freq)
 
     def test_get_freqs(self):
-        freq_dict = self.counter.get_freqs()
-        self.assertTrue("Filename", freq_dict.keys() - set(self.selected_measures))
+        freq_dict = self.counter1.get_freqs()
+        self.assertTrue("Filename", freq_dict.keys() - set(self.selected_measures1))