+++ CHANGELOG.md
@@ -1,3 +1,5 @@
++ Fix different result with and without `--select` (#12)
+
 <div align="center"><h1>Changelog</h1></div>
 
 ## [0.0.37](https://github.com/tanloong/neosca/releases/tag/0.0.37) (10 March 2023)