+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 <div align="center"><h1>Changelog</h1></div>
 
-## [0.0.38]() ()
+## [0.0.38](https://github.com/tanloong/neosca/releases/tag/0.0.38) (25 March 2023)
 
 ### New Features
 
@@ -10,7 +10,7 @@
 ### Bug Fixes
 
 + Fix not rasing error for typos in arguments when `--text` is specified
-+ Fix different result with and without `--select` (#12)
++ Fix zero frequency for some measures with `--select` (#12)
 
 ## [0.0.37](https://github.com/tanloong/neosca/releases/tag/0.0.37) (10 March 2023)